refac: create auth feature + create login form component
