+++ features/auth/index.tsx
@@ -1,0 +1,1 @@
+export * from "@/features/auth/components";
